+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/action/model/GetPropertyActionItem.java
@@ -56,11 +56,11 @@ public class GetPropertyActionItem extends ActionItem {
 
     @Override
     public GraphTextWriter appendTo(GraphTextWriter writer, LocalData localData) throws InterruptedException {
-        if (isEmptyString(target)) {
+        /*if (isEmptyString(target)) {
             return writer.append(Action.propertyNames[propertyIndex]);
-        }
+        }*/
 
-        if (!useGetPropertyFunction) {
+        if (!useGetPropertyFunction && !isEmptyString(target)) {
             target.appendToNoQuotes(writer, localData);
             writer.append(":");
             writer.append(Action.propertyNames[propertyIndex]);